+++ src/store/actions/index.js
@@ -1,1 +1,1 @@
-export * from "./auth"
+export * from "./actionCreator"
