+++ src/store/reducers/index.js
@@ -1,4 +1,4 @@
 import { combineReducers } from "redux";
-import auth from "./auth";
+import state from "./reducers";
 
-export default combineReducers({ auth });
+export default combineReducers({ state });
